pyproject.toml
[tool.poetry]
name = "sigmoidal"
version = "0.1.0"
description = ""
authors = ["Carlos Melo <carlos@sigmoidal.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.3,<3.12.0"
numpy = "^1.26.1"
scikit-learn = "^1.3.2"
supervision = "^0.16.0"
taskipy = "^1.12.0"
ipykernel = "^6.26.0"
pygments = "^2.16.1"
pymdown-extensions = "^10.3.1"
python-dotenv = "^1.0.0"
nbdev = "^2.3.13"
matplotlib = "^3.8.0"
nbformat = "^5.9.2"
ipytest = "^0.13.3"
imutils = "^0.5.4"
moviepy = "^1.0.3"
mediapipe = "^0.10.7"

[tool.poetry.dev-dependencies]
mkdocs = "^1.5.3"
isort = "^5.12.0"
blue = "^0.9.1"
mkdocstrings-python = "^1.7.3"
mkdocs-material = "^9.4.7"
pydocstyle = "^6.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_third_party = []

[tool.taskipy.tasks]
format = "isort . && blue ."
test = "pytest -v"
